moved home page to / instead of /home
